fix(backend): resolve max-lines regression by splitting upsert test file
Split assignmentDefinitionController.upsert.test.js (994 lines) into 4
test files + 1 shared helpers module, each under 500 lines:

- assignmentDefinitionUpsertTestHelpers.js (NEW - 117 lines): shared
  pure helper functions extracted from the original file
- assignmentDefinitionController.upsert.test.js (MODIFIED - 466 lines):
  20 core create/update, weighting, and save-validation tests
- assignmentDefinitionController.upsert.validation.test.js (NEW - 271
  lines): 10 duplicate identity, reparse, and task weighting tests
- assignmentDefinitionController.upsert.alternateTopics.test.js
  (NEW - 228 lines): 7 alternateTitles/alternateTopics flow tests
- assignmentDefinitionController.upsert.shape.test.js (NEW - 286
  lines): 7 canonical shape, yearGroupLabel, and re-parse tests

All 44 tests pass. Backend lint: clean (no max-lines violations on
any upsert file). 0 pre-existing warnings changed.
Regression checker: 0 regressions, 0 new failures, 1 fix.

ACTION_PLAN.md: update Section 10 implementation notes and mark
Section 1 technical debt as resolved.
